Deploy in Haroku host services

+++ node-complete-guide/util/file.js
@@ -2,11 +2,17 @@ const fs = require('fs');
 const { deleteOne } = require('../models/order');
 
 const deleteFile = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      throw (err);
+  try {
+    if(fs.existsSync(filePath)) {
+      fs.unlink(filePath, (err) => {
+        if (err) {
+          throw (err);
+        }
+      });
     }
-  });
+  } catch(err) {
+    throw (err);
+  }
 };
 
 exports.deleteFile = deleteFile;
